% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baplot-package.R
\docType{package}
\name{baplot}
\alias{baplot-package}
\alias{baplot}
\title{baplot: Bland–Altman plots with mixed-effects limits of agreement}
\description{
Package that launches a Shiny app to create Bland–Altman plots for method comparison studies,
}
\details{
Run runBA() to launch the Shiny app.usethis::use_vignette
}
\author{
\strong{Maintainer}: Jane Doe \email{jane@example.com}

}
\keyword{internal}
